requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-genai>=0.2.0
python-multipart>=0.0.6
pydantic>=2.9.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
oss2>=2.18.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0.0
pydub>=0.25.0

